+++ R/printers.R
@@ -13,6 +13,9 @@
 #' htmltools_value(flextable(iris[1:5,]))
 #' @importFrom htmltools tagList
 htmltools_value <- function(x, ft.align = "center"){
+
+  x <- flextable_global$defaults$post_process_html(x)
+
   html_o <- tagList(flextable_html_dependency(),
                     HTML(html_str(x, ft.align = ft.align, class = "tabwid",
                                   caption = caption_html_str(x, bookdown = FALSE),
@@ -45,8 +48,8 @@ htmltools_value <- function(x, ft.align = "center"){
 #' cell, the default value is 8 points.
 #' @param ft.arraystretch height of each row relative to its default
 #' height, the default value is 1.5.
-#' @param ft.left,ft.top Position should be defined with options \code{ft.left}
-#' and \code{ft.top}. Theses are the top left coordinates in inches
+#' @param ft.left,ft.top Position should be defined with options `ft.left`
+#' and `ft.top`. Theses are the top left coordinates in inches
 #' of the placeholder that will contain the table. Their
 #' default valuues are 1 and 2 inches.
 #' @param webshot webshot package as a scalar character, one of "webshot" or
@@ -161,6 +164,8 @@ flextable_to_rmd <- function(
 #' html_value(flextable(iris[1:5,]))
 html_value <- function(x, ft.align = opts_current$get("ft.align"), ft.shadow = opts_current$get("ft.shadow"), bookdown = FALSE, pandoc2 = TRUE){
 
+  x <- flextable_global$defaults$post_process_html(x)
+
   if(is.null(ft.shadow)){
     ft.shadow <- TRUE
   }
@@ -174,6 +179,7 @@ html_value <- function(x, ft.align = opts_current$get("ft.align"), ft.shadow = o
     if(pandoc2) "```{=html}",
     html_str(x, ft.align = ft.align, caption = caption_str, shadow = ft.shadow),
     if(pandoc2) "```",
+    "", "",
     sep = "\n")
   knit_meta_add(list(flextable_html_dependency()))
 
@@ -197,6 +203,8 @@ docx_value <- function(x,
                        ft.align = opts_current$get("ft.align"),
                        ft.split = opts_current$get("ft.split"),
                        bookdown = FALSE){
+
+  x <- flextable_global$defaults$post_process_docx(x)
 
   if( is.null(ft.align) ) ft.align <- "center"
   if( is.null(ft.split) ) ft.split <- FALSE
@@ -231,9 +239,10 @@ latex_value <- function(x,
                         ft.arraystretch = opts_current$get("ft.arraystretch"),
                         bookdown) {
   if (is.null(ft.align)) ft.align <- "center"
-  if (is.null(ft.tabcolsep)) ft.tabcolsep <- 8
+  if (is.null(ft.tabcolsep)) ft.tabcolsep <- 2
   if (is.null(ft.arraystretch)) ft.arraystretch <- 1.5
 
+  x <- flextable_global$defaults$post_process_pdf(x)
 
   fonts_ignore <- flextable_global$defaults$fonts_ignore
   fontspec_compat <- get_pdf_engine() %in% c("xelatex", "lualatex")
@@ -278,6 +287,7 @@ pptx_value <- function(x, ft.left = opts_current$get("ft.left"),
     ft.left <- 1
   if( is.null(ft.top) )
     ft.top <- 2
+  x <- flextable_global$defaults$post_process_pptx(x)
 
   uid <- as.integer(runif(n=1) * 10^9)
 
@@ -292,25 +302,25 @@ pptx_value <- function(x, ft.left = opts_current$get("ft.left"),
 #' @export
 #' @title flextable printing
 #'
-#' @description print a flextable object to format \code{html}, \code{docx},
-#' \code{pptx} or as text (not for display but for informative purpose).
+#' @description print a flextable object to format `html`, `docx`,
+#' `pptx` or as text (not for display but for informative purpose).
 #' This function is to be used in an interactive context.
 #'
 #' @note
-#' When argument \code{preview} is set to \code{"docx"} or \code{"pptx"}, an
+#' When argument `preview` is set to `"docx"` or `"pptx"`, an
 #' external client linked to these formats (Office is installed) is used to
 #' edit a document. The document is saved in the temporary directory of
 #' the R session and will be removed when R session will be ended.
 #'
-#' When argument \code{preview} is set to \code{"html"}, an
+#' When argument `preview` is set to `"html"`, an
 #' external client linked to these HTML format is used to display the table.
 #' If RStudio is used, the Viewer is used to display the table.
 #'
 #' Note also that a print method is used when flextable are used within
-#' R markdown documents. See \code{\link{knit_print.flextable}}.
+#' R markdown documents. See [knit_print.flextable()].
 #' @param x flextable object
 #' @param preview preview type, one of c("html", "pptx", "docx", "log").
-#' When \code{"log"} is used, a description of the flextable is printed.
+#' When `"log"` is used, a description of the flextable is printed.
 #' @param ... unused argument
 #' @family flextable print function
 #' @importFrom utils browseURL
@@ -361,7 +371,7 @@ print.flextable <- function(x, preview = "html", ...){
 #' minimum [pandoc](https://pandoc.org/installing.html) versions:
 #'
 #' \tabular{rc}{
-#'   \strong{Output format} \tab \strong{pandoc minimal version} \cr
+#'   **Output format** \tab **pandoc minimal version** \cr
 #'   HTML              \tab >= 1.12\cr
 #'   Word (docx)       \tab >= 2.0 \cr
 #'   PowerPoint (pptx) \tab >= 2.4 \cr
@@ -374,8 +384,9 @@ print.flextable <- function(x, preview = "html", ...){
 #' are to be used to change the default settings:
 #'
 #' \tabular{lcccccc}{
-#'   \strong{chunk option} \tab \strong{property} \tab \strong{default value} \tab \strong{HTML} \tab \strong{docx} \tab \strong{PDF} \tab \strong{pptx} \cr
+#'   **chunk option** \tab **property** \tab **default value** \tab **HTML** \tab **docx** \tab **PDF** \tab **pptx** \cr
 #'   ft.align        \tab flextable alignment, supported values are 'left', 'center' and 'right'    \tab 'center' \tab yes \tab yes \tab yes \tab no \cr
+#'   ft.shadow       \tab HTML option, disable shadow dom (set to `FALSE`) for pagedown. \tab TRUE    \tab yes  \tab no \tab no  \tab no \cr
 #'   ft.split        \tab Word option 'Allow row to break across pages' can be activated when TRUE. \tab FALSE    \tab no  \tab yes \tab no  \tab no \cr
 #'   ft.tabcolsep    \tab space between the text and the left/right border of its containing cell   \tab 8.0      \tab no  \tab no  \tab yes \tab no \cr
 #'   ft.arraystretch \tab height of each row relative to its default height                         \tab 1.5      \tab no  \tab no  \tab yes \tab no \cr
@@ -405,13 +416,13 @@ print.flextable <- function(x, preview = "html", ...){
 #' example. The table below expose these options:
 #'
 #' \tabular{llll}{
-#'   \strong{chunk option} \tab \strong{purpose} \tab \strong{rmarkdown} \tab \strong{bookdown} \cr
+#'   **chunk option** \tab **purpose** \tab **rmarkdown** \tab **bookdown** \cr
 #'   tab.cap.style \tab (Word only) style name to use for table captions            \tab yes \tab yes\cr
 #'   tab.cap.pre   \tab (Word only) Prefix for numbering chunk (default to "Table") \tab yes \tab yes\cr
 #'   tab.cap.sep   \tab (Word only) Suffix for numbering chunk (default to ": ")    \tab yes \tab yes\cr
-#'   tab.cap       \tab \strong{Caption label}                                      \tab yes \tab yes\cr
-#'   tab.id        \tab \strong{Caption reference unique identifier}                \tab yes \tab no \cr
-#'   label         \tab \strong{Caption reference unique identifier}                \tab no  \tab yes
+#'   tab.cap       \tab **Caption label**                                      \tab yes \tab yes\cr
+#'   tab.id        \tab **Caption reference unique identifier**                \tab yes \tab no \cr
+#'   label         \tab **Caption reference unique identifier**                \tab no  \tab yes
 #' }
 #' @section HTML output:
 #'
@@ -433,7 +444,7 @@ print.flextable <- function(x, preview = "html", ...){
 #'
 #' Also images cannot be integrated into tables with the PowerPoint format.
 #'
-#' @param x a \code{flextable} object
+#' @param x a `flextable` object
 #' @param ... further arguments, not used.
 #' @export
 #' @importFrom utils getFromNamespace
@@ -524,7 +535,7 @@ knit_print.flextable <- function(x, ...){
 #' @param ... flextable objects, objects, possibly named. If named objects, names are
 #' used as titles.
 #' @param values a list (possibly named), each element is a flextable object. If named objects, names are
-#' used as titles. If provided, argument \code{...} will be ignored.
+#' used as titles. If provided, argument `...` will be ignored.
 #' @param path HTML file to be created
 #' @param encoding encoding to be used in the HTML file
 #' @param title page title
@@ -559,6 +570,8 @@ save_as_html <- function(..., values = NULL, path, encoding = "utf-8", title = d
     if(show_names){
       txt[1] <- paste0("<h2>", titles[i], "</h2>")
     }
+    values[[i]] <- flextable_global$defaults$post_process_html(values[[i]])
+
     txt[2] <- html_str(values[[i]],
                        caption = caption_html_str(values[[i]], bookdown = FALSE),
                        shadow = FALSE)
@@ -587,7 +600,7 @@ save_as_html <- function(..., values = NULL, path, encoding = "utf-8", title = d
 #' @param ... flextable objects, objects, possibly named. If named objects, names are
 #' used as slide titles.
 #' @param values a list (possibly named), each element is a flextable object. If named objects, names are
-#' used as slide titles. If provided, argument \code{...} will be ignored.
+#' used as slide titles. If provided, argument `...` will be ignored.
 #' @param path PowerPoint file to be created
 #' @examples
 #' ft1 <- flextable( head( iris ) )
@@ -627,7 +640,7 @@ save_as_pptx <- function(..., values = NULL, path){
 #' @param ... flextable objects, objects, possibly named. If named objects, names are
 #' used as titles.
 #' @param values a list (possibly named), each element is a flextable object. If named objects, names are
-#' used as titles. If provided, argument \code{...} will be ignored.
+#' used as titles. If provided, argument `...` will be ignored.
 #' @param path Word file to be created
 #' @param pr_section a [prop_section] object that can be used to define page
 #' layout such as orientation, width and height.
@@ -699,7 +712,7 @@ save_as_docx <- function(..., values = NULL, path, pr_section = NULL){
 #' @note This function requires package webshot or webshot2.
 #' @param x a flextable object
 #' @param path image file to be created. It should end with .png, .pdf, or .jpeg.
-#' @param zoom,expand parameters used by \code{webshot} function.
+#' @param zoom,expand parameters used by `webshot` function.
 #' @param webshot webshot package as a scalar character, one of "webshot" or
 #' "webshot2".
 #' @examples
@@ -746,7 +759,7 @@ save_as_image <- function(x, path, zoom = 3, expand = 10, webshot = "webshot" ){
 #' result in a new R graphics window.
 #' @note This function requires packages: webshot and magick.
 #' @param x a flextable object
-#' @param zoom,expand parameters used by \code{webshot} function.
+#' @param zoom,expand parameters used by `webshot` function.
 #' @param ... additional parameters sent to [as_raster()] function
 #' @examples
 #' ftab <- flextable( head( mtcars ) )
@@ -771,7 +784,7 @@ plot.flextable <- function(x, zoom = 2, expand = 2, ... ){
 #' on a ggplot object.
 #' @note This function requires packages: webshot and magick.
 #' @param x a flextable object
-#' @param zoom,expand parameters used by \code{webshot} function.
+#' @param zoom,expand parameters used by `webshot` function.
 #' @param webshot webshot package as a scalar character, one of "webshot" or
 #' "webshot2".
 #' @importFrom grDevices as.raster
@@ -786,7 +799,7 @@ plot.flextable <- function(x, zoom = 2, expand = 2, ... ){
 #' @family flextable print function
 as_raster <- function(x, zoom = 2, expand = 2, webshot = "webshot"){
   if (!requireNamespace(webshot, quietly = TRUE)) {
-    stop("package webshot2 is required when saving a flextable as an image.")
+    stop("package ", webshot, " is required when saving a flextable as an image.")
   }
   if (!requireNamespace("magick", quietly = TRUE)) {
     stop("package magick is required when saving a flextable as an image.")