+++ R/tabpart_format.R
@@ -1,5 +1,5 @@
 # utils -----
-css_pt <- function(x, digits = 0){
+css_pt <- function(x, digits = 1){
   x <- ifelse( is.na(x), "inherit",
           ifelse( x < 0.001, "0",
                   paste0(format_double(x, digits = digits),"pt")))
@@ -49,7 +49,6 @@ border_pml <- function(color, width, style, side){
 # main -----
 
 #' @importFrom htmltools htmlEscape
-#' @importFrom gdtools raster_write raster_str
 #' @importFrom xml2 as_xml_document xml_find_all xml_attr
 format.complex_tabpart <- function( x, type = "wml", header = FALSE,
                                     split = FALSE, colwidth = TRUE, ... ){